+++ packages/camera/src/components/Layout/index.js
@@ -1,6 +1,6 @@
-import React, { useState } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
-import { ActivityIndicator, Platform, StyleSheet, View } from 'react-native';
+import { Platform, StyleSheet, View } from 'react-native';
 
 import useOrientation from '../../hooks/useOrientation';
 import useWindowDimensions from '../../hooks/useWindowDimensions';
@@ -23,15 +23,15 @@ const styles = StyleSheet.create({
 });
 
 function Layout({ children, containerStyle, left, right, sectionStyle }) {
-  const [isAvailable, setAvailable] = useState(false);
   const isNative = Platform.select({ native: true, default: false });
 
   const orientation = useOrientation();
   const { width: windowWidth } = useWindowDimensions();
-  const mobileBrowserIsPortrait = useMobileBrowserConfig(setAvailable);
+  const mobileBrowserIsPortrait = useMobileBrowserConfig();
 
-  if (!isAvailable) { return (<View />); }
-  if (mobileBrowserIsPortrait || (isNative && orientation.isNotLandscape)) { return (<View />); }
+  if (mobileBrowserIsPortrait || (isNative && orientation.isNotLandscape)) {
+    return (<View />);
+  }
 
   return (
     <View
@@ -51,7 +51,7 @@ function Layout({ children, containerStyle, left, right, sectionStyle }) {
           default: { maxWidth: 'calc(100% - 225px)' },
         }), sectionStyle]}
       >
-        {isAvailable ? children : <ActivityIndicator size="large" />}
+        {children}
       </View>
       <View
         accessibilityLabel="Side right"
